modularizes authSlice
