+++ src/store/index.js
@@ -1,24 +1,5 @@
 import { createSlice, configureStore } from '@reduxjs/toolkit';
 
-
-
-const initialAuthState = {
-    isAuthenticated: false,
-};
-
-const authSlice = createSlice({
-    name: 'auth',
-    initialState: initialAuthState,
-    reducers: {
-        login(state) {
-            state.isAuthenticated = true;
-        },
-        logout(state) {
-            state.isAuthenticated = false;
-        },
-    },
-});
-
 const store = configureStore({
     reducer: counterSlice.reducer,
     auth: authSlice.reducer,
